Create new zaddrs at run-time for new HushContacts
This shows how to generate a zaddr at run-time and use it correctly,
I am sure Denio will be happy! This method will work for Sietch in SDL,
but actually I need to throw all this code away for this exact use case.
At least Denio can use it!

What I did was to study how the code to generate new zaddrs in the Receive
UI works, and this is the TLDR: It uses a chain of callbacks to assure that
class-level data it stores (the RPC::zaddresses variable) always has data
when it goes to read it. I did the lazy thing and pushed everything into
a single callback, which is equivalent.

The core problem now is that a brand new zaddr will not have any funds,
and is unable to pay the default tx fee. We must always make sure we have a
pool of these addresses, at startup, and make sure they have some funds,
which can be periodically checked and "topped up". Once we have that,
we will pull a new address from that pool if we add a new HushContact, instead
of this current run-time generation of a new zaddr.